requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-login==0.6.3
flask-wtf==1.2.1
flask-oauthlib==0.9.6
flask-mail==0.9.1
authlib==1.2.1
anthropic
Pillow==10.1.0
python-dotenv==1.0.0
stripe==7.6.0
pymysql==1.1.0
cryptography==41.0.5
werkzeug==2.3.7
gunicorn==21.2.0